api/requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
brotli-python==1.1.0
bzip2==1.0.8
c-ares==1.22.1
ca-certificates==2023.11.17
cached-property==1.5.2
cached_property==1.5.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
exceptiongroup==1.2.0
fastapi==0.104.1
flatbuffers==23.5.26
freetype==2.12.1
frozenlist==1.4.0
gast==0.5.4
giflib==5.2.1
google-auth-oauthlib==1.0.0
google-auth==2.23.4
google-pasta==0.2.0
grpcio==1.54.3
h11==0.14.0
h5py==3.10.0
hdf5==1.14.2
icu==73.2
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
keras==2.14.0
krb5==1.21.2
lcms2==2.15
lerc==4.0.0
libabseil==20230125.3
libaec==1.1.2
libblas==3.9.0
libcblas==3.9.0
libcurl==8.4.0
libcxx==16.0.6
libdeflate==1.19
libedit==3.1.20191231
libev==4.33
libffi==3.4.2
libgfortran5==13.2.0
libgfortran==5.0.0
libgrpc==1.54.3
libjpeg-turbo==3.0.0
liblapack==3.9.0
libnghttp2==1.58.0
libopenblas==0.3.25
libpng==1.6.39
libprotobuf==3.21.12
libsqlite==3.44.1
libssh2==1.11.0
libtiff==4.6.0
libwebp-base==1.3.2
libxcb==1.15
libzlib==1.2.13
llvm-openmp==17.0.5
markdown==3.5.1
markupsafe==2.1.3
ml_dtypes==0.2.0
multidict==6.0.4
ncurses==6.4
numpy==1.26.0
oauthlib==3.2.2
openjpeg==2.5.0
openssl==3.1.4
opt_einsum==3.3.0
packaging==23.2
pillow==10.1.0
pip==23.3.1
protobuf==4.21.12
pthread-stubs==0.4
pyasn1-modules==0.3.0
pyasn1==0.5.0
pycparser==2.21
pydantic-core==2.14.5
pydantic==2.5.2
pyjwt==2.8.0
pyopenssl==23.3.0
pysocks==1.7.1
python-flatbuffers==23.5.26
python-multipart==0.0.6
python==3.10.13
python_abi==3.10
pyu2f==0.1.5
re2==2023.03.02
readline==8.2
requests-oauthlib==1.3.1
requests==2.31.0
rsa==4.9
setuptools==68.2.2
six==1.16.0
snappy==1.1.10
sniffio==1.3.0
starlette==0.27.0
tensorboard-data-server==0.7.0
tensorboard==2.14.1
tensorflow-base==2.14.0
tensorflow-estimator==2.14.0
tensorflow==2.14.0
termcolor==2.3.0
tk==8.6.13
typing-extensions==4.8.0
typing_extensions==4.8.0
tzdata==2023c
urllib3==2.1.0
uvicorn==0.24.0
werkzeug==3.0.1
wheel==0.41.3
wrapt==1.14.1
xorg-libxau==1.0.11
xorg-libxdmcp==1.1.3
xz==5.2.6
yarl==1.9.2
zipp==3.17.0
zlib==1.2.13
zstd==1.5.5
